backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pandas
numpy
python-dotenv
pydantic-settings